WebSecurityConfig: add security configs that the h2-console works again
Disable csrf, we don't want the hustle to send a CRSF token back
and forth.
Add contentSecurityPolicy (CSP) to allow the h2-console to submit forms to this server under the specified domains.
Add frameOptions to allow frames to display content of the
spring application. Else the browser restricts displaying the h2 console,
which seems to be build with old frame technology.