GameRepository: only allow members of a game to change the game
According to the tutorial on:
https://www.baeldung.com/spring-security-create-new-custom-security-expression

Show it with the test in FollowUpGameControllerIntegrationTest that
a player who is not part of the game cannot create the next game.

The idea is that you can implement the BelongsToGame interface on every entity,
and then use belongsToGame(#entity.game) in the @PreAuthorize expression.

Issue: #154
Issue: #140